% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{construct.arglist}
\alias{construct.arglist}
\title{Evaluate all function arguments}
\usage{
construct.arglist(funobj, framepos = -1)
}
\arguments{
\item{funobj}{the function whose arguments are to be evaluated}

\item{framepos}{TBD}
}
\value{
named list containing function arguments
}
\description{
Evaluates all function arguments at location where construct.arglist is called in that function
(they might have changed during running the function, or may be in variables)
construct.arglist gets rid of the variables and so on, and returns the argument list as a list
}
