% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_threshold}
\alias{get_threshold}
\title{Sample the latent thresholds through griddy Gibbs}
\usage{
get_threshold(Achg1, SIGMAS1, SIGMAS2, threshgrid, Aapprox)
}
\arguments{
\item{Achg1}{TODO}

\item{SIGMAS1}{TODO}

\item{SIGMAS2}{TODO}

\item{threshgrid}{TODO}

\item{Aapprox}{TODO}
}
\description{
This function implements the griddy Gibbs sampler for the thresholds
}
