% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_function.R
\name{MCMC_tvp}
\alias{MCMC_tvp}
\title{Run MCMC for the TTVP model}
\usage{
MCMC_tvp(Y, X, nburn, nsave, priorbtheta = list(B_1 = 2, B_2 = 1, kappa0
  = 1e-07), priorb0 = list(a_tau = 0.1, c_tau = 0.01, d_tau = 0.01),
  priormu = c(0, 10^2), priorsig = c(0.01, 0.01), priorphi = c(2, 2),
  h0prior = "stationary", grid.length = 150, thrsh.pct = 0.1,
  thrsh.pct.high = 1, sv_on = TRUE, TVS = TRUE, cons.mod = TRUE,
  nr, thin = 0.1, robust = TRUE, a.approx = FALSE,
  sim.kappa = FALSE, kappa.grid = seq(1e-04, 0.1, 10))
}
\description{
some story
}
